Made animation titles editable
